requirements.txt
simple-salesforce==1.12.5
pandas==2.1.4
rapidfuzz==3.6.1
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
openpyxl==3.1.2
colorama==0.4.6
tqdm==4.66.1
